zaawansowany/urls.py
"""zaawansowany URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/dev/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.conf import settings
from django.contrib import admin
from django.urls import path
from biblioteka import views as v1
from django.contrib.auth.views import (PasswordResetView, PasswordResetDoneView, PasswordResetConfirmView,
                                       PasswordResetCompleteView)


urlpatterns = [
    path('admin/', admin.site.urls),
    path('', v1.index),
    path('email/', v1.email),
    path('password_reset/', PasswordResetView.as_view(), name="password_reset"),
    path('password_reset_done/',PasswordResetDoneView.as_view(), name="password_reset_done"),
    path('password_reset_confirm/<uidb64>/<token>',PasswordResetConfirmView.as_view(), name="password_reset_confirm"),
    path('password_reset_complete/',PasswordResetCompleteView.as_view(), name="password_reset_complete"),
]

if settings.DEBUG:
    from debug_toolbar.toolbar import debug_toolbar_urls
    urlpatterns += debug_toolbar_urls()
