requirements.txt
asgiref==3.8.1
Django==5.1.1
pytz==2019.3
sqlparse==0.4.4
django-debug-toolbar~=4.4.6